Search JUVIX_LLVM_DIST_PATH for the clang executable
JUVIX_LLVM_DIST_PATH should point to the root of a LLVM installation,
i.e clang should be present in $JUVIX_LLVM_DIST_PATH/bin/clang.

This will give us flexibility if we need to use other programs present
in the LLVM bin directory.

If JUVIX_LLVM_DIST_PATH is not set, or clang is not available there then
the system PATH is consulted instead.