Propogate formatter results when traversing subdirs
Previously if a project looked like:

Dir
|- juvix.yaml
|- Unformatted.juvix
|- Subdir
   |- Formatted.juvix

and the user ran `juvix format Dir/` then the command would return exit
code 0. This is because only the result from formatting files within
`Subdir` was used, the result from `Dir` was discarded.

This PR fixes this, the results of formatting all subdirectories in a
project are combined appropriately.