Support `juvix format` with no argument to format a project
When `juvix format` is invoked from some directory within a juvix
project then the formatter is run on all the files contained in the
project.

If `juvix format` is run from some directory outside of a Juvix project
then an error is reported. The user gets the same error as they would get if
`juvix format` was run with a directory argument that is not within a
Juvix project.