Version the PackageDescription format
The idea is that new versions of the PackageDescription format will be
added as files of the form:
include/package/PackageDescription/Vx.juvix

The correspondence between a version of the PackageDescription file and the package type name is recorded in `packageDescriptionTypes`.

Now the `package` identifier type must come from **one** of the versions
of the PackageDescription module.