Add PackageDescription.Basic module Package variant
This PR adds the PackageDescription.Basic module, available to
Package.juvix files.

```
module Package;

import PackageDescription.Basic open;

package : Package := basicPackage;
```

The PackageDescription.Basic module provides a Package type that is
translated to a Juvix Package with all default arguments. It is not
possible to customize a Basic package.

A Basic package does not depend on the standard library.