Set the Package file project root to the root of the loaded project
The Package loader PathResolver must the root of the project being loaded
to the Package.juvix file.

The order of clauses in the match changed by this commit must not be
permuted because the global 'package' package now contains its own
Package.juvix file. Previously a project's Package.juvix was incorrectly
being assigned to the global project root.